mock config refactor
