+++ ChartCareMVC/MVC-Tests/TestSetup/MockConfig.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Mvc;
+using Moq;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.AspNetCore.Mvc.Abstractions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,5 +12,34 @@ namespace MVC_Tests.TestSetup
 {
     internal class MockConfig
     {
+        public Mock<IUrlHelper> CreateMockUrlHelper(ActionContext? context = null)
+        {
+            context ??= GetActionContextForPage("/Page");
+
+            var urlHelper = new Mock<IUrlHelper>();
+            urlHelper.SetupGet(h => h.ActionContext)
+                .Returns(context);
+            return urlHelper;
+        }
+        protected static ActionContext GetActionContextForPage(string page)
+        {
+            return new ActionContext
+            {
+                ActionDescriptor = new ActionDescriptor
+                {
+                    RouteValues = new Dictionary<string, string?>
+            {
+                { "page", page },
+            }
+                },
+                RouteData = new RouteData
+                {
+                    Values =
+            {
+                ["page"] = page
+            }
+                }
+            };
+        }
     }
 }
